% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_it.R
\name{batch_it}
\alias{batch_it}
\title{Batch a vector or list into a list of elements with a maximum size}
\usage{
batch_it(x, batch_size)
}
\arguments{
\item{x}{A vector or list}

\item{batch_size}{numeric. The size (length) of batches to create. Should be
a single positive integer value (see Examples).}
}
\description{
Batch a vector or list into a list of elements with a maximum size
}
\examples{
# ----
batch_it(letters, 6L)
batch_it(letters, 27L)

}
