% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{exclude_codes}
\alias{exclude_codes}
\title{Provides a vector of fields to return, which excludes all ONS code fields}
\usage{
exclude_codes()
}
\description{
Provides a vector of fields to return, which excludes all ONS code fields
}
