% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{filter_fields}
\alias{filter_fields}
\title{Helper function to check a vector of fields to filter}
\usage{
filter_fields(fields = NULL)
}
\arguments{
\item{fields}{character vector. See \link{schema_table}.}
}
\description{
Helper function to check a vector of fields to filter
}
