% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_fixes.R
\name{fix_by_autocomplete}
\alias{fix_by_autocomplete}
\title{Use the autocompletion API to suggest replacements for invalid postcodes}
\usage{
fix_by_autocomplete(codes)
}
\description{
Use the autocompletion API to suggest replacements for invalid postcodes
}
\keyword{internal}
