% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_fixes.R
\name{fix_terminated}
\alias{fix_terminated}
\title{Use reverse geocoding to suggest replacements for terminated postcodes}
\usage{
fix_terminated(dat)
}
\description{
Use reverse geocoding to suggest replacements for terminated postcodes
}
\keyword{internal}
