% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_postcode_data.R
\name{get_postcode_data}
\alias{get_postcode_data}
\title{Use postcodes.io to get postcode data}
\usage{
get_postcode_data(x, as_list = FALSE, include_codes = TRUE)
}
\arguments{
\item{x}{A vector of postcodes}

\item{as_list}{boolean, default \code{FALSE}. The default behaviour is to
extract the data from the list and return it as a tibble. If set \code{TRUE},
an unprocessed list of the JSON data from the API will be returned instead}

\item{include_codes}{boolean, default \code{TRUE}. Include columns for the ONS
administrative codes for geographic units in the returned tibble. Irrelevant
if \code{as_list} is \code{TRUE}; in this case all ONS codes are included as a
nested list in the list data for each postcode.
See \url{https://postcodes.io/docs/postcode/lookup} for details}
}
\value{
A tibble by default, otherwise a list if \code{as_list} is TRUE
}
\description{
See \url{https://postcodes.io/docs/postcode/lookup} for details of the data
that the API returns
}
\examples{
get_postcode_data(c("NP22 3PS", "NP22 4PS", "NP22 5PS"))
}
