% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHSRpostcodetools-package.R
\docType{package}
\name{NHSRpostcodetools-package}
\alias{NHSRpostcodetools}
\alias{NHSRpostcodetools-package}
\title{NHSRpostcodetools: Get geographic and administrative data about UK postcodes}
\description{
The main purpose of the package is to extract data from the postcodes.io API for collections of UK postcodes, such as their LSOA (or equivalent), parliamentary constituency, and spatial coordinates. It can also check postcodes for currency/validity and suggest replacement current postcodes for terminated codes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nhs-r-community.github.io/NHSRpostcodetools/}
  \item \url{https://github.com/nhs-r-community/NHSRpostcodetools}
  \item Report bugs at \url{https://github.com/nhs-r-community/NHSRpostcodetools/issues}
}

}
\author{
\strong{Maintainer}: Fran Barton \email{francis.barton@nhs.net} (\href{https://orcid.org/0000-0002-5650-1176}{ORCID})

Authors:
\itemize{
  \item Zoë Turner \email{zoe.turner3@nhs.net} (\href{https://orcid.org/0000-0003-1033-9158}{ORCID})
}

Other contributors:
\itemize{
  \item NHS-R Community \email{nhs.rcommunity@nhs.net} [copyright holder]
}

}
\keyword{internal}
