% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_postcode_data.R
\name{postcode_data_join}
\alias{postcode_data_join}
\title{Look up data for postcodes from a data frame column and join the results on}
\usage{
postcode_data_join(tbl, .col = "postcode", filter = filter_fields())
}
\arguments{
\item{tbl}{A data frame}

\item{.col}{string. The name of the column that contains the postcodes.
\code{"postcode"} by default.}

\item{filter}{character vector or helper function. The default is
\link{filter_fields}, which by default evaluates to \code{NULL}, which means no
fields will be filtered out from the returned data. You can also use
\link{filter_fields} to validate a vector of field names that you supply - see
Examples below.
The full list of field names is available in the \link{schema_table} data. Some
other helper functions are provided for common scenarios: \link{exclude_codes}
and \link{minimal_fields}.
See \url{https://postcodes.io/docs/postcode/bulk} for details.}
}
\description{
Look up data for postcodes from a data frame column and join the results on
}
\examples{
tibble::tibble(
  place = paste0("place_", seq(3L)),
  postcode = c("NP22 3PS", "NP22 4PS", "NP22 5PS")
) |>
  postcode_data_join()
}
