% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_geocode.R
\name{reverse_geocode}
\alias{reverse_geocode}
\title{Recursive function to use the reverse geocoding API to find current
postcodes to replace terminated codes, by proximity}
\usage{
reverse_geocode(dat, prev_data = NULL, curr_radius = 125L)
}
\description{
Recursive function to use the reverse geocoding API to find current
postcodes to replace terminated codes, by proximity
}
\keyword{internal}
