% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_table.R
\docType{data}
\name{schema_table}
\alias{schema_table}
\title{Copy of the postcodes.io schema summary table}
\format{
\subsection{\code{schema_table}}{

A data frame with 7,240 rows and 60 columns:
\describe{
\item{field}{field name}
\item{type}{type of data expected for this field}
\item{description}{description of the field and its format}
\item{example}{an example of the type of data expected to be returned}
}
}
}
\source{
\url{https://postcodes.io/docs/postcode/schema}
}
\usage{
schema_table
}
\description{
Copy of the postcodes.io schema summary table
}
\keyword{datasets}
