% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{tidy_postcodes}
\alias{tidy_postcodes}
\title{Try to format a list of postcodes to the "A(A)D(D) DAA" PCD8 format}
\usage{
tidy_postcodes(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A character vector
}
\description{
Try to format a list of postcodes to the "A(A)D(D) DAA" PCD8 format
}
