% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_postcode_data.R
\name{unnest_codes}
\alias{unnest_codes}
\title{Unnest codes (wider): from a list-col to a column each}
\usage{
unnest_codes(tbl)
}
\arguments{
\item{tbl}{A data frame with ONS (etc.) codes data in a list-col}
}
\description{
Unnest codes (wider): from a list-col to a column each
}
\keyword{internal}
