% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_fixes.R
\name{zero_row}
\alias{zero_row}
\title{A convenience function to generate a 0-row tibble to support suggest_fixes()}
\usage{
zero_row()
}
\description{
A convenience function to generate a 0-row tibble to support suggest_fixes()
}
\keyword{internal}
