% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHSRpostcodetools-package.R
\docType{package}
\name{NHSRpostcodetools-package}
\alias{NHSRpostcodetools}
\alias{NHSRpostcodetools-package}
\title{NHSRpostcodetools: Get geographic and administrative data about UK postcodes}
\description{
Extracts data from the postcodes.io API for a given UK postcode, such as its LSOA (or equivalent), parliamentary constituency, and spatial coordinates. It can also check a list of postcodes for currency/validity and suggest a replacement current postcode for an expired code.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nhs-r-community.github.io/NHSRpostcodetools/}
  \item \url{https://github.com/nhs-r-community/NHSRpostcodetools}
  \item Report bugs at \url{https://github.com/nhs-r-community/NHSRpostcodetools/issues}
}

}
\author{
\strong{Maintainer}: Zoë Turner \email{zoe.turner3@nhs.net} (\href{https://orcid.org/0000-0003-1033-9158}{ORCID})

Authors:
\itemize{
  \item Fran Barton \email{francis.barton@nhs.net} (\href{https://orcid.org/0000-0002-5650-1176}{ORCID})
}

Other contributors:
\itemize{
  \item NHS-R community \email{nhs.rcommunity@nhs.net} [copyright holder]
}

}
\keyword{internal}
