% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_postcode_data.R
\name{postcode_data_join}
\alias{postcode_data_join}
\title{Look up data for postcodes from a data frame column and join the results on}
\usage{
postcode_data_join(tbl, .col = "postcode", include_codes = TRUE)
}
\arguments{
\item{tbl}{A data frame}

\item{.col}{string. The name of the column that contains the postcodes.
\code{"postcode"} by default.}

\item{include_codes}{boolean, default \code{TRUE}. Include columns for the ONS
administrative codes for geographic units in the returned tibble. Irrelevant
if \code{as_list} is \code{TRUE}; in this case all ONS codes are included as a
nested list in the list data for each postcode.
See \url{https://postcodes.io/docs/postcode/lookup} for details}
}
\description{
Look up data for postcodes from a data frame column and join the results on
}
\examples{
tibble::tibble(
  place = paste0("place_", seq(3L)),
  postcode = c("NP22 3PS", "NP22 4PS", "NP22 5PS")
) |>
  postcode_data_join()
}
