% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_fixes.R
\name{suggest_fixes}
\alias{suggest_fixes}
\title{Return data for suggested replacements for invalid postcodes}
\usage{
suggest_fixes(codes)
}
\arguments{
\item{codes}{A character vector of postcodes to be checked for validity}
}
\value{
A data frame with a row for each invalid postcode supplied.
}
\description{
Data for terminated invalid codes will be returned according to the nearest
current valid code for the old code's longitude/latitude, if possible.
Data for codes that are invalid due to an incorrect final letter will be
returned according to an autocompletion process, if possible.
Missing data will be returned for any other invalid postcodes that could not
be replaced using the two methods above.
Any valid codes will be excluded from the returned data frame.
}
\examples{
suggest_fixes(c("hd1 2ut", "hd1 2uu", "hd1 2uv"))
}
